Move source dist building into its own module
